[ch11864] update name file lang.js to store.js and refactoring of store.js about naming and replication in others files
