--- src/js/lang.js
@@ -1,11 +1,0 @@
-var store =  {
-    debug: true,
-    state: {
-        lang: ''
-    },
-    setState (value) {
-        this.state.lang = value;
-    }
-};
-
-module.exports = store;
