+++ webpack.config.js
@@ -15,7 +15,7 @@ if (process.env.NODE_ENV === 'production') {
 }
 
 module.exports = {
-    entry: './src/js/main.js',
+    entry: './src/main.js',
     output: {
         path: path.resolve(__dirname, './dist'),
         publicPath: '/',
